+++ source/R/plot.gv.R
@@ -1,31 +1,57 @@
 plot.gv <-
 function(x, line.res=100, pch=1,
-         legend=TRUE, leg.x=NA, leg.y=NA, leg.cex=1, ...) {
-    X <- x$lag[!is.na(x$lag)]
+         legend=TRUE, leg.x=NA, leg.y=NA, leg.cex=1,
+         bar.length=0.1, bar.col="gray", bar.lty=par("lty"), ...) {
+### TODO: check if it is a multi should be via a class attribut
+### or in the object parameters slot.
+### plot.gv can hadle multi and single gv objects
+
+    ## check if gv is multi
+    multi <- FALSE
+    if ("gamma.mat" %in% names(x)) multi <- TRUE
+
+    ## Get all distance classes to X-axis
+    X <- x$lag[!is.na(x$lag)] 
     Y <- x$gamma[!is.na(x$gamma)]
 
+    # Check if a model is present
     mtest <- mtest.gv(x)
     if (mtest) {
         xx <- seq(0, ceiling(max(X)), length.out = line.res)
         yy <- predict(x, xx)
     }
 
-    #Prepare point size relative to n
+    # Prepare point size relative to n
     lab.n <- range(x$n)
     cex <- 2 * x$n / lab.n[2]
     
-    # Begin plot
+    ## ### Begin plotting ### ##
     plot.new()
+
+    ## Get X- and Y- axes ranges
     x.range <- c(0, max(X))
+    mY <- max(Y)
+    if (multi) mY <- max(x$gamma.ci[2,])
     if (mtest) {
-        y.range <- c(0, max(c(Y,yy)))
+        y.range <- c(0, max(c(mY,yy)))
     } else {
-        y.range <- c(0, max(Y))
+        y.range <- c(0, mY)
     }
 
     plot.window(x.range, y.range)
-    points(X,Y, pch=pch, cex=cex*leg.cex, ...)
+    if (multi) {
+        ##lines(X, x$gamma.ci[1,], lty=2)
+        ##lines(X, x$gamma.ci[2,], lty=2)
+        ##lines(X, Y, lty=1)
+        ## Using bars instead of lines
+        arrows(X, x$gamma.ci[1,], X, x$gamma.ci[2,], angle = 90,
+               length = bar.length, code = 3, col = bar.col,
+               lty = bar.lty)
+        
+    }
     
+    points(X,Y, pch=pch, cex=cex*leg.cex, ...)
+
     if (mtest) lines(xx, yy, col='red', ...)
 
     axis(1)