+++ source/R/predict.gv.R
@@ -13,36 +13,43 @@ function(object, newdata, ...) {
     N <- object$model$nugget
     R <- object$model$range    
 
-    if (model == 1) {
-    # Gaussian model
-        vFUN <- function(x) N + (S-N) * (1-exp(-3*(x**2)/(R**2)))
-
-    } else if (model ==2 ) {
-    # Exponential mode
-        vFUN <- function(x) N + (S-N) * (1-exp(-x/(R)))
-
-    } else if (model == 3) {
-    # Spherical model
-        vFUN <- function(x) {
-            y <- x * 0 + S
-            y[x<R] <- N+(S-N)*(((3*(x[x<R]))/(2*R))-((x[x<R])**3/(2*R**3)))
-            return(y)
+    if (model == "gaussian") {
+        vFUN <- function(x, S, N, R) {
+            N + (S-N) * (1-exp(-3*(x**2)/(R**2)))
         }
-    } else if (model == 4) {
-        Sl <- object$model$slope
-        if (is.infinite(S)) {
-            # Linear model
-            vFUN <- function(x) N + Sl*x
-        } else {
-            # Linear model with sill
-            vFUN <- function(x) {
-                y <- x * 0 + S
-                y[x<R] <- N + Sl * x[x<R]
-                return(y)
-            }
-        }            
+
+    } else if (model == "exponential") {
+        vFUN <- function(x, S, N, R) {
+            N + (S-N) * (1-exp(-x/(R)))
+        }
+        
+    } else if (model == "spherical") {
+        vFUN <- function(x, S, N, R) {
+            ifelse(x<=R,
+                   N+(S-N)*(((3*(x))/(2*R))-((x)**3/(2*R**3))),
+                   S)
+        }
+        
+    } else if (model == "pentaspherical") {
+        vFUN <- function(x, S, N, R) {
+            ifelse(x<=R, 
+                   N+(S-N)*(((15/8)*(x/R)) - ((5/4)*(x/R)**3) + ((3/8)*(x/R)**5)),
+                   S)
+        }
+        
+    } else if (model == "cubic") {
+        vFUN <- function(x, S, N, R) {
+            ifelse(x<=R,
+                   N+(S-N)*(7*(x/R)**2 - (35/4)*(x/R)**3 + (7/2)*(x/R)**5 - (3/4)*(x/R)**7),
+                   S)
+        }
+        
+    } else if (model == "linear") {
+        vFUN <- function(x, S, N, R) {
+            ifelse(x<=R, ((S-N)/R)*x+N, S)
+        }
     }
 
-    y <- vFUN(newdata)
+    y <- vFUN(newdata, S, N, R)
     y
 }