+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Improve distance calculations with resistance/friction information using 'gdistance' package.	
 * The "krig" function was modified to accept a distance function to calculate distances. The 'default' uses simple euclidean distances between coordinates to maintain compatibility with code for previous versions. 
+* Plot method for 'gv' was duplicated and is now corrected.
+* Plotting gv with multiple trees now displays CI bars instead of lines.	
 	
 # Version 1.1.1